requirements.txt
asgiref==3.8.1
config==0.5.1
Django==5.0.6
django-admin-rangefilter==0.13.0
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
